+++ src/Shared/AWS/AWSSemanticConventions.v1.34.0.cs
@@ -1,0 +1,26 @@
+// Copyright The OpenTelemetry Authors
+// SPDX-License-Identifier: Apache-2.0
+
+namespace OpenTelemetry.AWS;
+
+// disable Style Warnings to improve readability of this specific file.
+#pragma warning disable SA1124
+#pragma warning disable SA1005
+#pragma warning disable SA1514
+#pragma warning disable SA1201
+#pragma warning disable SA1516
+
+internal partial class AWSSemanticConventions
+{
+    /// <summary>
+    /// Open Telemetry Semantic Conventions as of 1.34.0:
+    /// https://github.com/open-telemetry/semantic-conventions/releases/tag/v1.34.0.
+    /// </summary>
+    private class AWSSemanticConventions_V1_34_0 : AWSSemanticConventions_V1_29_0
+    {
+        // AWS Attributes
+        public override string AttributeAWSSNSTopicArn => "aws.sns.topic.arn";
+
+        public override string AttributeAWSSQSQueueUrl => "aws.sqs.queue.url";
+    }
+}