+++ src/OpenTelemetry.Instrumentation.AWS/CHANGELOG.md
@@ -8,6 +8,9 @@
 * Add the `aws.s3.bucket` and `aws.s3.key` attributes to S3 spans.
   ([#4029](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4029))
 
+* Capture SNS `TopicArn` as the `aws.sns.topic.arn` span attribute.
+  ([#4043](https://github.com/open-telemetry/opentelemetry-dotnet-contrib/pull/4043))
+
 ## 1.15.0
 
 Released 2026-Jan-21